+++ integration/proxy_rolling_test.go
@@ -97,8 +97,8 @@ func (it *ProxyRollingTestSuite) TestRollingThreats() {
 		src           string
 		dst           string
 		fqdn          string
-		finalScore    float32
-		beaconScore   float32
+		finalScore    float64
+		beaconScore   float64
 		totalDuration float64
 		totalBytes    float64
 		count         uint64
